+++ ataskq/test_ataskq.py
@@ -1,10 +1,19 @@
 from pathlib import Path
 from datetime import datetime, timedelta
+from copy import copy
 
 import pytest
 
 from ataskq import TaskQ, StateKWArg, Job, Task, targs, EStatus
 from ataskq.db_handler import DBHandler
+from ataskq.handler import EAction
+
+from .tasks_utils import dummy_args_task
+
+
+@pytest.fixture
+def jtaskq(conn) -> TaskQ:
+    return TaskQ(conn=conn).create_job()
 
 
 def test_create_job(conn):
@@ -30,6 +39,222 @@ def test_job_default_name(conn):
 def test_job_custom_name(conn):
     job = TaskQ(conn=conn).create_job(name='my_job').job
     assert job.name == 'my_job'
+
+
+def test_update_task_start_time(jtaskq):
+    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+
+    jtaskq.add_tasks(
+        [
+            in_task1,
+        ]
+    )
+
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 1  # sanity
+    task: Task = tasks[0]
+
+    # update db with task time (also updates task inplace)
+    start_time = datetime.now()
+    start_time = start_time.replace(microsecond=0)  # test to seconds resolution
+    jtaskq.update_task_start_time(task, start_time)
+    assert task.start_time == start_time
+
+    # get task form db and validate
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 1  # sanity
+    task: Task = tasks[0]
+    assert task.start_time == start_time
+
+
+def test_update_task_status(jtaskq):
+    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+
+    jtaskq.add_tasks(
+        [
+            in_task1,
+        ]
+    )
+
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 1  # sanity
+    task: Task = tasks[0]
+    assert task.status == EStatus.PENDING
+
+    now = datetime.now()
+    # update db with task status (also updates task inplace)
+    jtaskq.update_task_status(task, EStatus.RUNNING, timestamp=now)
+    assert task.status == EStatus.RUNNING
+    assert task.pulse_time == now
+    assert task.done_time is None
+    # get task form db and validate
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 1  # sanity
+    task: Task = tasks[0]
+    assert task.status == EStatus.RUNNING
+    assert task.pulse_time == now
+    assert task.done_time is None
+
+    # update db with task status (also updates task inplace)
+    jtaskq.update_task_status(task, EStatus.SUCCESS, timestamp=now)
+    assert task.status == EStatus.SUCCESS
+    assert task.pulse_time == now
+    assert task.done_time == now
+    # get task form db and validate
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 1  # sanity
+    task: Task = tasks[0]
+    assert task.status == EStatus.SUCCESS
+    assert task.pulse_time == now
+    assert task.done_time == now
+
+
+def test_take_next_task_sanity(jtaskq):
+    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+
+    jtaskq.add_tasks(
+        [
+            in_task1,
+        ]
+    )
+
+    action, task = jtaskq._take_next_task(level=None)
+    assert action == EAction.RUN_TASK
+    _compare_task_taken(in_task1, task)
+
+
+def _compare_task_taken(task1: Task, task2: Task, job_id=None):
+    if job_id:
+        assert task1.job_id == job_id
+    assert task1.job_id == task2.job_id
+    assert task1.name == task2.name
+    assert task1.description == task2.description
+    assert task1.level == task2.level
+    assert task1.entrypoint == task2.entrypoint
+    assert task1.targs == task2.targs
+
+
+def test_take_next_task(jtaskq):
+    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+    in_task2 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, name="task2")
+    in_task3 = copy(in_task1)
+
+    jtaskq.add_tasks(
+        [
+            in_task2,
+            in_task1,
+            in_task3,
+        ]
+    )
+
+    tids = []
+
+    action, task = jtaskq._take_next_task(level=None)
+    assert action == EAction.RUN_TASK
+    assert task.task_id not in tids
+    tids.append(task.task_id)
+    _compare_task_taken(in_task1, task)
+
+    action, task = jtaskq._take_next_task(level=None)
+    assert action == EAction.RUN_TASK
+    assert task.task_id not in tids
+    tids.append(task.task_id)
+    _compare_task_taken(in_task3, task)  # note in_task3 is copy of in_task1
+
+    action, task = jtaskq._take_next_task(level=None)
+    assert action == EAction.WAIT
+    assert task is None
+
+
+def test_take_next_task_2_jobs(conn):
+    # todo: test should ne under ataskq
+    taskq1: TaskQ = TaskQ(conn=conn).create_job(name="job1")
+    taskq2: TaskQ = TaskQ(conn=conn).create_job(name="job2")
+
+    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, name="taska")
+    in_task2 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="taskb")
+    in_task3 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="taskb")
+    in_task4 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="taskd")
+    in_task5 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, name="taske")
+
+    taskq1.add_tasks(
+        [
+            in_task2,
+            in_task1,
+            in_task3,
+        ]
+    )
+
+    taskq2.add_tasks(
+        [
+            in_task5,
+            in_task4,
+        ]
+    )
+
+    jid1 = taskq1.job_id
+    jid2 = taskq2.job_id
+    tids = []
+
+    # sanity check
+    assert len(Job.get_all(_handler=taskq1.handler)) == 2
+    assert len(taskq1.get_tasks()) == 3
+    assert len(Job.get_all(_handler=taskq1.handler)) == 2
+    assert len(taskq2.get_tasks()) == 2
+
+    # db handler 1
+    action, task = taskq1._take_next_task(level=None)
+    assert action == EAction.RUN_TASK
+    assert task.task_id not in tids
+    tids.append(task.task_id)
+    _compare_task_taken(in_task2, task, job_id=jid1)
+    tids.append(task.task_id)
+
+    action, task = taskq1._take_next_task(level=None)
+    assert action == EAction.RUN_TASK
+    assert task.task_id not in tids
+    tids.append(task.task_id)
+    _compare_task_taken(in_task3, task, job_id=jid1)  # note in_task3 is copy of in_task1
+
+    action, task = taskq1._take_next_task(level=None)
+    assert action == EAction.WAIT
+    assert task is None
+
+    # db handler 2
+    action, task = taskq2._take_next_task(level=None)
+    assert action == EAction.RUN_TASK
+    assert task.task_id not in tids
+    tids.append(task.task_id)
+    _compare_task_taken(in_task4, task, job_id=jid2)
+    tids.append(task.task_id)
+
+    action, task = taskq2._take_next_task(level=None)
+    assert action == EAction.WAIT
+    assert task is None
+
+
+def test_get_tasks(jtaskq):
+    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
+    in_task2 = Task(entrypoint=dummy_args_task, level=2, name="task2")
+    in_task3 = Task(entrypoint=dummy_args_task, level=3, name="task3")
+
+    jtaskq.add_tasks(
+        [
+            in_task3,
+            in_task2,
+            in_task1,
+        ]
+    )
+
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 3
+    for t in tasks:
+        if t.level == 1:
+            _compare_task_taken(in_task1, t)
+        elif t.level == 2:
+            _compare_task_taken(in_task2, t)
+        elif t.level == 3:
+            _compare_task_taken(in_task3, t)
 
 
 def test_run_default(conn, tmp_path: Path):
